refactor: move scroll state into native state
